+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 Improvements:
 * Core: Completely re-invented the blocks system; it's now position-driven.
 * Core: removed has_extra and extra_ids from pages database and replaced it with joins resulting in the same result but based upon real data (rather than just relying on the existing scripts.)
+* Installer: added 'getTemplateId' function to easily fetch a template id.
 
 
 2.6.3 (xxxx-xx-xx)